# app-demo3/Model/Empleado.cs
namespace app_demo3.Model
{
    public class Empleado
    {
        public String? id { get; set; }
        public String? nombre { get; set; }
        public String? apellido { get; set; }
        public String? cargo { get; set; }
        public DateTime fecha_contrato { get; set; }
    }
}