# app-demo3/Model/Producto.cs
namespace app_demo3.Model
{
    public class Producto
    {
        public String? id { get; set; }
        public String? nombre { get; set; }
        public String? descripcion { get; set; }
        public float? precio { get; set; }
        public String? categoria_id { get; set; }
        public int? stock { get; set; }

    }
}